Application released
This is the point where the application was released to the public.

I built this at the start of my career as a developer, and coming back
to it found it impossible to make head or tail of the narrative left in
my git commit messages, so I have squashed them all.