A second round of refactoring
This commit is a second round of refactoring based on having been a
professional developer for a year.

It removes a lot of the logic from the submissions model, it also
converts submissions data from a large relational mess of submissions
answers and targets (80+ rows per submission) to use JSON data fields
held on submission. This has the added advantage of bringing the
application back down to below the row limit for the free Heroku tier.

There is also much improved test coverage.