Add progress bar when creating a new submission.
This commit adds a progress bar when a user is creating a new
submission. The form is very long and it is helpful to know where you
are.

The progress bar is only shown for new submissions, as users may jump
between areas on existing submissions and may not be going in order.

This was requested as a feature request on github:
https://github.com/timpwbaker/digital-maturity-matrix/issues/1